tidy up document.title setter
